refactor cause data creation
